Create de.json
This is a first translation if the interface and certain parts in German. We intend to continue it in the following days and expect it to be completed soon. This would need a thorough review afterwards, we may also consider changing the wording in some cases. We would love to coordinate the German translation, though. 
Please contact us if you have any further questions.